pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openclaw-node"
version = "0.1.0"
description = "Remote execution framework for VFX/post-production pipeline automation"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "OpenClaw Team" }
]
keywords = ["vfx", "pipeline", "automation", "websocket"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.26",
    "ruff>=0.2",
    "mypy>=1.8",
]
shotgrid = [
    "shotgun-api3>=3.4",
]

[project.scripts]
openclaw-node = "client.main:main_cli"
openclaw-server = "server.main:main_cli"

[tool.hatch.build.targets.wheel]
packages = ["server", "client", "plugins"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
